+++ Samples/IonicCapacitorAngularV2/ct-demo-ionic-angular-cap/src/app/services/data.service.ts
@@ -24,7 +24,12 @@ export class DataService {
         "Open Inbox", "Show Total Counts", "Show Unread Counts", "Get All Inbox Messages",
         "Get Unread Messages", "Get InboxMessage by messageID", "Delete InboxMessage by messageID",
         "Mark as read by messageID",
-        "Notification Viewed event for Message", "Notification Clicked event for Message"
+        "Notification Viewed event for Message", "Notification Clicked event for Message",
+        "Fetch Inbox"
+      ]
+      ],
+      ["NATIVE DISPLAY", [
+        "Push Display Unit Element Clicked"
       ]
       ],
       ["PUSH TEMPLATES", [