+++ src/android/CleverTapPlugin.java
@@ -58,6 +58,7 @@ import com.clevertap.android.sdk.InAppNotificationListener;
 import com.clevertap.android.sdk.events.EventDetail;
 import com.clevertap.android.sdk.UTMDetail;
 import com.clevertap.android.sdk.CTInboxListener;
+import com.clevertap.android.sdk.FetchInboxCallback;
 import com.clevertap.android.sdk.CTInboxStyleConfig;
 import com.clevertap.android.sdk.inbox.CTInboxMessage;
 import com.clevertap.android.sdk.InboxMessageButtonListener;
@@ -107,7 +108,7 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         cleverTap.registerPushPermissionNotificationResponseListener(this);
 
         String libName = "Cordova";
-        int libVersion = 50000;
+        int libVersion = 50100;
         cleverTap.setLibrary(libName);
         cleverTap.setCustomSdkVersion(libName, libVersion);
 
@@ -1015,6 +1016,21 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
         });
     }
 
+    private void fetchInbox(CallbackContext callbackContext) {
+        cordova.getThreadPool().execute(() -> cleverTap.fetchInbox(isSuccess -> sendPluginResult(callbackContext, Status.OK, isSuccess)));
+    }
+
+    private void pushDisplayUnitElementClickedEventForID(JSONArray args, CallbackContext callbackContext) {
+        executeWithArgs(args, callbackContext, (arguments) -> {
+            final String unitId = arguments.getString(0);
+            final JSONObject jsonProps = arguments.getJSONObject(1);
+            final HashMap<String, Object> additionalProperties = toMap(jsonProps);
+            cordova.getThreadPool().execute(() -> {
+                cleverTap.pushDisplayUnitElementClickedEventForID(unitId, additionalProperties);
+                sendPluginResult(callbackContext, Status.NO_RESULT);
+            });
+        });
+    }
 
     private void isFeatureFlagInitialized(CallbackContext callbackContext) {
         cordova.getThreadPool().execute(() -> {
@@ -1916,6 +1932,12 @@ public class CleverTapPlugin extends CordovaPlugin implements SyncListener, InAp
 
             case UNMUTE:
                 unmute(callbackContext);
+                return true;
+            case FETCH_INBOX:
+                fetchInbox(callbackContext);
+                return true;
+            case PUSH_DISPLAY_UNIT_ELEMENT_CLICKED_EVENT_FOR_ID:
+                pushDisplayUnitElementClickedEventForID(args, callbackContext);
                 return true;
             default: {
                 sendPluginResult(callbackContext, Status.ERROR, "unhandled CleverTapPlugin action");